archive/requirements.txt
bar-chart-race==0.1.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
httplib2==0.31.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.3.5
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psutil==7.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.6.2
